Update the TODO file.
Add some flags when we are being bitchy.

Some misc clean ups, converting things to size_t.

Signed->unsigned clean ups.

Kill many of the random unions scattered through the engine which just
happen to almost exactly match float_int_t.

Removed some fill variables in the model loader, as I'm no longer
searching for smashing. (Why did I commit those?)

New types, Sint and Uint, signed and unsigned int.

<qualifier> <type> instead of <type> <qualifier> (IE, inline void
instead of void inline)
(This is due to C standards.)

Move things around in gl_alias.c, optimize a little more.

Clean up the float->ub conversion with mod function slightly.

Optimize out some unneeded bits in the lightmap code.

Remove an unused menu option or two.

Kill some of the nastier bits of the NQ network code.

Rewrite the time tracking code to deal with reality better.
(Kills twibugs_qwsv_time_jump)

Don't draw spectators from the QW scoreboard in the minihud.

Kill random unused/unneeded cruft.

Clean up Q_sqrt.

Kill unused math funcs that just mirror the #defines.

glTexParameterf -> glTexParameteri

+++ trunk/twilight/src/base/crc.c
@@ -93,7 +93,7 @@ CRC_Value (unsigned short crcvalue)
 }
 
 unsigned short
-CRC_Block (Uint8 *start, int count)
+CRC_Block (Uint8 *start, size_t count)
 {
 	unsigned short crc;
 
@@ -183,7 +183,7 @@ For proxy protecting
 ====================
 */
 Uint8
-COM_BlockSequenceCRCByte (Uint8 *base, int length, int sequence)
+COM_BlockSequenceCRCByte (Uint8 *base, size_t length, int sequence)
 {
 	unsigned short crc;
 	Uint8      *p;
